works on featured league card
